rename: 폴더 및 파일명 변경| fix: 화살표 함수로 통일
