+++ src/components/MajorCard/index.tsx
